fix: Add get_pmc_for_date method and fix type conversions (8 errors fixed)
Added missing get_pmc_for_date method to TrainingAnalysisService to fix E0599 errors in feature_engineering_service.rs.

Fixed type mismatch errors by converting f64 PMC values to f32 for TrainingFeatures struct.

Progress: 222 → 216 errors remaining (8 E0599 and E0308 errors resolved)

+++ ai-coach-api/src/services/feature_engineering_service.rs
@@ -61,9 +61,9 @@ impl FeatureEngineeringService {
         let days_until_goal_event = self.get_days_until_goal_event(user_id).await?;
 
         Ok(TrainingFeatures {
-            current_ctl: pmc.ctl,
-            current_atl: pmc.atl,
-            current_tsb: pmc.tsb,
+            current_ctl: pmc.ctl as f32,
+            current_atl: pmc.atl as f32,
+            current_tsb: pmc.tsb as f32,
             days_since_last_workout,
             avg_weekly_tss_4weeks,
             recent_performance_trend,
@@ -130,9 +130,9 @@ impl FeatureEngineeringService {
         let days_until_goal_event = self.get_days_until_goal_event(user_id).await?;
 
         Ok(TrainingFeatures {
-            current_ctl: pmc.ctl,
-            current_atl: pmc.atl,
-            current_tsb: pmc.tsb,
+            current_ctl: pmc.ctl as f32,
+            current_atl: pmc.atl as f32,
+            current_tsb: pmc.tsb as f32,
             days_since_last_workout,
             avg_weekly_tss_4weeks,
             recent_performance_trend,
